remove question service and add sql translator
